+++ Toolkit/RSPiX/GREEN/BLiT/Fspr1.cpp
@@ -317,7 +317,7 @@ int16_t ConvertToFSPR1(RImage* pImage)
 		}
 
 	memcpy(pNewCodeBuf,pCodeBuf,lCompressedSize);
-	if (long(pCode - pCodeBuf + 1) > lMaxSize)
+   if (int32_t(pCode - pCodeBuf + 1) > lMaxSize)
 		{
 		TRACE("ConvertToFSPR1: I overran my own buffer!\n");
 		}
@@ -1062,7 +1062,7 @@ int16_t rspBlit(
 	afrSkipY = rspfruint16_tStrafe256(sDstH,sDenY);
 	// Make magnification possible:
 	int16_t i;
-	long *alDstSkip = (long*)calloc(sizeof(int32_t),afrSkipY[1].mod + 2);
+   int32_t *alDstSkip = (int32_t*)calloc(sizeof(int32_t),afrSkipY[1].mod + 2);
 	for (i=1;i<(afrSkipY[1].mod + 2);i++) 
 		alDstSkip[i] = alDstSkip[i-1] + lDstP;
 
@@ -1538,7 +1538,7 @@ int16_t rspBlit(
 	afrSkipY = rspfruint16_tStrafe256(sDstH,sDenY);
 	// Make magnification possible:
 	int16_t i;
-	long *alDstSkip = (long*)calloc(sizeof(int32_t),afrSkipY[1].mod + 2);
+   int32_t *alDstSkip = (int32_t*)calloc(sizeof(int32_t),afrSkipY[1].mod + 2);
 	for (i=1;i<(afrSkipY[1].mod + 2);i++) 
 		alDstSkip[i] = alDstSkip[i-1] + lDstP;
